hice un millon de cosas, ni me acuerdo :/

+++ TP3-EDA-FlightSim/main.cpp
@@ -1,56 +1,69 @@
 /************************ LIBRERIAS ************************/
 #include <stdio.h>
 #include <stdlib.h>
+#include <new>
 #include <stdbool.h>
 #include <time.h>
 #include "Random.h"
 #include "parser.h"
+#include "Callback.h"
 #include "UserData.h"
 #include "Tweety.h"
 #include "Drawing.h"
 
+#define HEIGHT 70
+#define WIDTH 100
+
 int main(int argc, char* argv[])
 {
 	randomize();					//Seed para nros random
-	init_all();							//Inicializo allegro
-	userData_t* myinfo;
-	tweety* birds;
-	bool quit = false;
-	char whatsNext;				//Variable con la informacion de la event_queue
-	if (parseCmdLine(argc, argv, /*&parseCallBack*/, &myinfo))
+
+	if (init_all())							//Inicializo allegro
 	{
-		//ME PARECE QUE ES ASI: new tweety[getUserData(myinfo,BIRDS)];
-		birds = (tweety*) malloc(sizeof(tweety) * myinfo->birdCount);
-		//create_birds();														//no se si hay alguna funcion en tweety.cpp para hacer los pajaritos
-		while (!quit)
+		userData myinfo;
+
+		bool quit = false;
+		char whatsNext;				//Variable con la informacion de la event_queue
+
+		if (parseCmdLine(argc, argv, parseCallBack, &myinfo))
 		{
-			whatsNext = itstime();
-			switch (whatsNext)
+			tweety* flock = new (std::nothrow) tweety[myinfo.getBirdCount()];
+
+			//randomize segun modo (falta modo)
+
+			while (!quit)
 			{
-				case ITSTIME:
-									for (int i = 0; i < myinfo->birdCount; i++)
-									{
-										//birds[i].proyect(birds);
-									}
-									for (int i = 0; i < myinfo->birdCount; i++)
-									{
-										//birds[i].move(birds);
-									}
-									//draw_birds(b);
-									//tick_count++;
-									break;
-				case '1':case '2':case 'E':case 'J':case 'V': case UP:case DOWN:
-									modify(whatsNext, myinfo, birds);
-									break;
-				case DO_EXIT:
-									quit = true;
-									break;
-				default:			break;
+				whatsNext = itstime();
+				switch (whatsNext)
+				{
+					case ITSTIME:
+										for (int i = 0; i < myinfo.getBirdCount(); i++)
+										{
+											flock[i].project(flock, myinfo.getBirdCount(), myinfo.getRandomJiggleLimit(), myinfo.getEyesight());
+										}
+										for (int i = 0; i < myinfo.getBirdCount(); i++)
+										{
+											flock[i].move(HEIGHT, WIDTH);
+										}
+										draw_birds(&myinfo, flock);
+										//tick_count++;
+										break;
+
+					case '1':case '2':case 'E':case 'J':case 'V': case UP:case DOWN:
+										modify(whatsNext, &myinfo, flock);
+										break;
+					case DO_EXIT:
+										quit = true;
+										break;
+
+					default:			break;
+				}
 			}
 		}
+		destroy_all();
+
 	}
-	destroy_all();
-	destroyUserData(myinfo);
+
 	return 1;
 }
